BZ2Compressor and BZ2Decompressor live in the _bz2 module
